Making t9 works
The converter for json+ld files is async and require a callback a function to work. Luckily we just needed to work around it with promises in the test.helper.
T11, 12, 13 have problems with BlankNodes ids. That's why an operational kb separated from the main one would be a good optional parameter for the fetcher. Also the xmlns:str is treated differently in the lib and the red.